Demo and support add image node.
A image defined by the url to a local file taint the canvas
and the screenshot funtionality don't work.
This commit add to 2_features.html the logic to add a image node,
and add support to process the dataurl in the canves.
The image is scaled to reduce the data size.