Add basic support to show images inside the nodes
In the example, the image is defined gby a url,
when the image is defined by the url to a local file taint the canvas
and the screenshot funtionality don't work.

This commit add to 2_features.html the logic to add a image node,
and add support to process the dataurl in the canves.
The image is scaled to reduce the data size.

ant.png image is from http://www.art4apps.org/
and is released under a Creative Commons Attribution-NonCommercial-ShareAlike license.
(CC BY-NC-SA)